+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Budiansah 2025 DOTATATE ([doi:10.1186/s40658-025-00726-7](https://doi.org/10.1186/s40658-025-00726-7)) -- adults with neuroendocrine tumours or meningiomas undergoing peptide receptor radionuclide therapy dosimetry.
+- Add Warren 2025 apremilast and orismilast ([doi:10.1007/s13555-025-01371-9](https://doi.org/10.1007/s13555-025-01371-9)) -- adults with atopic dermatitis, plus an in vitro human whole-blood IL-13 inhibition layer for each drug.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.